backend/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
google-api-core==1.22.1
google-auth==1.20.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.0
pytesseract==0.3.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
